* 출석이벤트 작업중

+++ resources/assets/js/components/Item.vue
@@ -31,23 +31,26 @@
                     <label class="col-sm-3 control-label"><span class="hidden-sm">출석 </span>이벤트</label>
                     <div class="col-sm-9">
                         <div class="input-group">
-                            <input type="number" min="1" max="300" class="form-control">
+                            <input type="number" min="1" max="300" class="form-control" :readonly="!this.enableOpen" v-model="item.event_rank">
                             <span class="input-group-btn">
-                                <button class="btn btn-sm btn-success"><span class="hidden-sm">이벤트 </span>시작</button>
-                                <button class="btn btn-sm btn-default">당첨자 확인</button>
+                                <button class="btn btn-sm btn-success" :disabled="!this.enableOpen" @click="eventOpen">
+                                    <span class="hidden-sm">이벤트 </span>시작
+                                </button>
+                                <button class="btn btn-sm btn-default" :disabled="!this.enableResult" @click="eventResult">
+                                    당첨자 확인
+                                </button>
                             </span>
                         </div>
                     </div>
                 </div>
             </div>
             <div class="panel-footer text-right">
-                <button class="btn btn-sm btn-info" v-on:click="edit">수정</button>
-                <button class="btn btn-sm btn-danger" v-on:click="remove">삭제</button>
+                <button class="btn btn-sm btn-info" @click="edit" :disabled="this.$parent.notReady">수정</button>
+                <button class="btn btn-sm btn-danger" @click="remove" :disabled="this.$parent.notReady">삭제</button>
             </div>
         </div>
     </div>
 </template>
-
 
 <script>
     export default {
@@ -67,6 +70,14 @@
                description : null
             }
         },
+        computed: {
+            enableOpen: function () {
+                return this.$parent.state == 'open' && !this.item.event_open;
+            },
+            enableResult: function () {
+                return this.$parent.state == 'open' && this.item.event_open;
+            }
+        },
         methods: {
             edit: function(event) {
                 axios.patch('admin/items/'+this.item.id, this.item)
@@ -78,14 +89,36 @@
                 });
             },
             remove: function(event) {
-                axios.delete('admin/items/'+this.item.id)
-                .then((response) => {
-                    alert('삭제 되었습니다.');
-                    this.$emit('remove');
-                })
-                .catch((error) => {
-                    console.log(error);
-                });
+                if (window.confirm('['+this.item.title+'] 삭제 하시겠습니까?')) {
+                    axios.delete('admin/items/'+this.item.id)
+                    .then((response) => {
+                        this.$emit('remove');
+                    })
+                    .catch((error) => {
+                        console.log(error);
+                    });
+                }
+            },
+            eventOpen: function(event) {
+                if (window.confirm('출석 이벤트를 시작 하시겠습니까?')) {
+                    axios.patch('admin/items/'+this.item.id+'/event', {"open": true, "rank":this.item.event_rank})
+                    .then((response) => {
+                        this.item.event_open = true;
+                        this.item.event_winner = null;
+                    })
+                    .catch((error) => {
+                        console.log(error);
+                    });
+                }
+            },
+            eventResult: function(event) {
+                axios.patch('admin/items/'+this.item.id+'/event', {"open": false})
+                    .then((response) => {
+                        this.item.event_open = false;
+                    })
+                    .catch((error) => {
+                        console.log(error);
+                    });
             }
         }
     }
